Move invoke to separate file, rename Stream, add interface.
