+++ grpcweb/client.go
@@ -92,83 +92,3 @@ func (c Client) RPCCall(ctx context.Context, method string, req []byte, opts ...
 		return nil, ctx.Err()
 	}
 }
-
-// Stream performs a server-to-client streaming RPC call, returning
-// a struct which exposes a Go gRPC like streaming interface.
-// It is non-blocking.
-func (c Client) Stream(ctx context.Context, method string, req []byte, opts ...CallOption) (*StreamClient, error) {
-	srv := &StreamClient{
-		ctx:      ctx,
-		messages: make(chan []byte, 10), // Buffer up to 10 messages
-		errors:   make(chan error, 1),
-	}
-
-	onMsg := func(in []byte) { srv.messages <- in }
-	onEnd := func(s *status.Status) {
-		if s.Code != codes.OK {
-			srv.errors <- s
-		} else {
-			srv.errors <- io.EOF
-		}
-	}
-	err := invoke(ctx, c.host, c.service, method, req, onMsg, onEnd, opts...)
-	if err != nil {
-		return nil, err
-	}
-
-	return srv, nil
-}
-
-// Invoke populates the necessary JS structures and performs the gRPC-web call.
-// It attempts to catch any JS errors thrown.
-func invoke(ctx context.Context, host, service, method string, req []byte, onMsg onMessageFunc, onEnd onEndFunc, opts ...CallOption) (err error) {
-	methodDesc := newMethodDescriptor(newService(service), method, newResponseType())
-
-	c := &callInfo{}
-	rawOnEnd := func(code int, msg string, trailers metadata.Metadata) {
-		s := &status.Status{
-			Code:     codes.Code(code),
-			Message:  msg,
-			Trailers: trailers.MD,
-		}
-		c.trailers = trailers.MD
-
-		// Perform CallOptions required after call
-		for _, o := range opts {
-			o.after(c)
-		}
-
-		onEnd(s)
-	}
-	onHeaders := func(headers metadata.Metadata) {
-		c.headers = headers.MD
-	}
-
-	md, _ := gmd.FromOutgoingContext(ctx)
-	props := newProperties(host, false, newRequest(req), metadata.New(md), onHeaders, onMsg, rawOnEnd)
-
-	// Recover any thrown JS errors
-	defer func() {
-		e := recover()
-		if e == nil {
-			return
-		}
-
-		if e, ok := e.(*js.Error); ok {
-			err = e
-		} else {
-			panic(e)
-		}
-	}()
-
-	// Perform CallOptions required before call
-	for _, o := range opts {
-		if err := o.before(c); err != nil {
-			return status.FromError(err)
-		}
-	}
-
-	js.Global.Get("grpc").Call("invoke", methodDesc, props)
-
-	return nil
-}
